requirements.txt
aiogram==2.14.3
python-dotenv==0.19.0
psutil==5.8.0
requests==2.26.0